refactor: extract automation cli handlers
